Initial commit
This is all awful. Please do not use it. I may make it better later,
but right now it's a giant hack around not necessarily needing protobuf
support but still needing to use protobuf message definitions to
structure messages in languages where a full protobuf compiler would be
a serious undertaking. i.e., this is the lazy approach.