Move basic template functions into formatting.go
Also adds case-conv.go, which just contains a few rudimentary functions
for doing camel/snake/pascal case strings. The assumption for
pascal/camel case is that the string was previously snake case or
separated by spaces. Snake case allows you to specify a separator,
which can be a string of any length.